split operator Hono API route graph
