+++ templates/operator/src/entry.ts
@@ -5,6 +5,7 @@ import type {
   createInlineDispatcher,
   StepDispatcher,
 } from "@voyantjs/workflows-orchestrator-cloudflare"
+import { dispatchHonoApiRequest, isHonoApiRequest } from "./hono-api-dispatch"
 import {
   CHANNEL_PUSH_AVAILABILITY_CRON,
   CHANNEL_PUSH_BOOKING_LINK_CRON,
@@ -14,12 +15,6 @@ import {
 } from "./scheduled-crons"
 
 const startHandler = createStartHandler(defaultStreamHandler)
-
-let apiAppPromise: Promise<typeof import("./api/app")["app"]> | undefined
-function loadApiApp(): Promise<typeof import("./api/app")["app"]> {
-  apiAppPromise ??= import("./api/app").then((mod) => mod.app)
-  return apiAppPromise
-}
 
 let workflowDefinitionsPromise: Promise<unknown> | undefined
 function loadWorkflowDefinitions(): Promise<unknown> {
@@ -44,13 +39,8 @@ export default {
     const url = new URL(request.url)
 
     // Route /api/* to Hono (strip prefix so Hono sees /v1/*, /auth/*, /health)
-    if (url.pathname === "/api" || url.pathname.startsWith("/api/")) {
-      const stripped = url.pathname.slice(4) || "/"
-      const apiUrl = new URL(stripped, url.origin)
-      apiUrl.search = url.search
-      const apiRequest = new Request(apiUrl.toString(), request)
-      const apiApp = await loadApiApp()
-      return apiApp.fetch(apiRequest, env, ctx)
+    if (isHonoApiRequest(url.pathname)) {
+      return dispatchHonoApiRequest(request, env, ctx)
     }
 
     // Everything else → TanStack Start SSR
